switched name and added 'add_subscription'
